+++ src/main/java/org/jetbrains/bsp/bazel/server/bazel/BUILD
@@ -6,9 +6,9 @@ java_library(
     visibility = ["//visibility:public"],
     deps = [
         "//src/main/java/org/jetbrains/bsp/bazel/common",
-        "//src/main/java/org/jetbrains/bsp/bazel/server/bazel/data:bazel-data",
-        "//src/main/java/org/jetbrains/bsp/bazel/server/bazel/parameters:bazel-parameters",
-        "//src/main/java/org/jetbrains/bsp/bazel/server/bazel/utils:bazel-utils",
+        "//src/main/java/org/jetbrains/bsp/bazel/server/bazel/data",
+        "//src/main/java/org/jetbrains/bsp/bazel/server/bazel/params",
+        "//src/main/java/org/jetbrains/bsp/bazel/server/bazel/utils",
         "@io_bazel//src/main/protobuf:analysis_java_proto",
         "@io_bazel//src/main/protobuf:build_java_proto",
         "@maven//:ch_epfl_scala_bsp4j",
