+++ src/main/java/org/jetbrains/bsp/bazel/server/bazel/BazelRunnerBuilder.java
@@ -4,8 +4,8 @@ import com.google.common.collect.ImmutableList;
 import java.util.ArrayList;
 import java.util.List;
 import org.jetbrains.bsp.bazel.server.bazel.data.BazelProcessResult;
-import org.jetbrains.bsp.bazel.server.bazel.parameters.BazelQueryKindParameters;
-import org.jetbrains.bsp.bazel.server.bazel.parameters.BazelRunnerFlag;
+import org.jetbrains.bsp.bazel.server.bazel.params.BazelQueryKindParameters;
+import org.jetbrains.bsp.bazel.server.bazel.params.BazelRunnerFlag;
 import org.jetbrains.bsp.bazel.server.bazel.utils.BazelArgumentsUtils;
 
 public class BazelRunnerBuilder {
