+++ src/main/java/org/jetbrains/bsp/bazel/server/bazel/params/BUILD
@@ -1,7 +1,7 @@
 load("@rules_java//java:defs.bzl", "java_library")
 
 java_library(
-    name = "bazel-parameters",
+    name = "params",
     srcs = glob(["*.java"]),
     visibility = ["//visibility:public"],
     deps = [
