+++ src/main/java/org/jetbrains/bsp/bazel/server/bazel/params/BazelQueryKindParameters.java
@@ -1,4 +1,4 @@
-package org.jetbrains.bsp.bazel.server.bazel.parameters;
+package org.jetbrains.bsp.bazel.server.bazel.params;
 
 public class BazelQueryKindParameters {
