+++ src/main/java/org/jetbrains/bsp/bazel/server/bazel/utils/BUILD
@@ -1,11 +1,11 @@
 load("@rules_java//java:defs.bzl", "java_library")
 
 java_library(
-    name = "bazel-utils",
+    name = "utils",
     srcs = glob(["*.java"]),
     visibility = ["//visibility:public"],
     deps = [
-        "//src/main/java/org/jetbrains/bsp/bazel/server/bazel/parameters:bazel-parameters",
+        "//src/main/java/org/jetbrains/bsp/bazel/server/bazel/params",
         "@maven//:ch_epfl_scala_bsp4j",
         "@maven//:com_google_guava_guava",
     ],
