+++ src/main/java/org/jetbrains/bsp/bazel/server/bsp/BUILD
@@ -7,8 +7,8 @@ java_library(
     deps = [
         "//src/main/java/org/jetbrains/bsp/bazel/common",
         "//src/main/java/org/jetbrains/bsp/bazel/server/bazel",
-        "//src/main/java/org/jetbrains/bsp/bazel/server/bazel/data:bazel-data",
-        "//src/main/java/org/jetbrains/bsp/bazel/server/bazel/parameters:bazel-parameters",
+        "//src/main/java/org/jetbrains/bsp/bazel/server/bazel/data",
+        "//src/main/java/org/jetbrains/bsp/bazel/server/bazel/params",
         "//src/main/java/org/jetbrains/bsp/bazel/server/bep",
         "//src/main/java/org/jetbrains/bsp/bazel/server/resolver",
         "//src/main/java/org/jetbrains/bsp/bazel/server/util",
