+++ src/main/java/org/jetbrains/bsp/bazel/server/bsp/BazelBspServerBuildManager.java
@@ -34,8 +34,8 @@ import org.jetbrains.bsp.bazel.common.Uri;
 import org.jetbrains.bsp.bazel.server.bazel.BazelRunner;
 import org.jetbrains.bsp.bazel.server.bazel.data.BazelData;
 import org.jetbrains.bsp.bazel.server.bazel.data.BazelProcessResult;
-import org.jetbrains.bsp.bazel.server.bazel.parameters.BazelQueryKindParameters;
-import org.jetbrains.bsp.bazel.server.bazel.parameters.BazelRunnerFlag;
+import org.jetbrains.bsp.bazel.server.bazel.params.BazelQueryKindParameters;
+import org.jetbrains.bsp.bazel.server.bazel.params.BazelRunnerFlag;
 import org.jetbrains.bsp.bazel.server.bep.BepServer;
 import org.jetbrains.bsp.bazel.server.resolver.QueryResolver;
 import org.jetbrains.bsp.bazel.server.util.ParsingUtils;
