+++ src/main/java/org/jetbrains/bsp/bazel/server/resolver/ActionGraphResolver.java
@@ -6,7 +6,7 @@ import java.io.IOException;
 import java.util.List;
 import org.jetbrains.bsp.bazel.server.bazel.BazelRunner;
 import org.jetbrains.bsp.bazel.server.bazel.data.BazelProcessResult;
-import org.jetbrains.bsp.bazel.server.bazel.parameters.BazelRunnerFlag;
+import org.jetbrains.bsp.bazel.server.bazel.params.BazelRunnerFlag;
 import org.jetbrains.bsp.bazel.server.util.ActionGraphParser;
 
 public class ActionGraphResolver {
