+++ src/main/java/org/jetbrains/bsp/bazel/server/service/BuildServerService.java
@@ -44,8 +44,8 @@ import org.jetbrains.bsp.bazel.common.Uri;
 import org.jetbrains.bsp.bazel.server.bazel.BazelRunner;
 import org.jetbrains.bsp.bazel.server.bazel.data.BazelData;
 import org.jetbrains.bsp.bazel.server.bazel.data.BazelProcessResult;
-import org.jetbrains.bsp.bazel.server.bazel.parameters.BazelQueryKindParameters;
-import org.jetbrains.bsp.bazel.server.bazel.parameters.BazelRunnerFlag;
+import org.jetbrains.bsp.bazel.server.bazel.params.BazelQueryKindParameters;
+import org.jetbrains.bsp.bazel.server.bazel.params.BazelRunnerFlag;
 import org.jetbrains.bsp.bazel.server.bsp.BazelBspServerBuildManager;
 import org.jetbrains.bsp.bazel.server.bsp.BazelBspServerLifetime;
 import org.jetbrains.bsp.bazel.server.bsp.BazelBspServerRequestHelpers;
